pyproject.toml
[project]
name = "svg-wallpaper-gen"
version = "0.1.0"
description = ""
authors = [
    {name = "Gwenn",email = "gwenn.garrigues@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=12.0.0,<13.0.0)",
    "marimo (>=0.18.4,<0.19.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[tool.poetry]
packages = [{include = "svg_wallpaper_gen", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.marimo.ai]
rules = "- prefer polars over pandas\n- make charts using altair"

[tool.marimo.save]
autosave = "after_delay"
autosave_delay = 1000
format_on_save = true

[tool.marimo.package_management]
manager = "poetry"